# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.VocabularyReviewRequest:
    properties:
      quality:
        description: 0 -> 5
        example: 4
        maximum: 5
        minimum: 0
        type: integer
      vocabulary_id:
        example: 22222222-2222-2222-2222-222222222222
        format: uuid
        type: string
    type: object
  domain.VocabularyReviewResponse:
    properties:
      interval_days:
        example: 1
        type: integer
      new_stage:
        example: 1
        type: integer
      next_review_at:
        type: string
      vocabulary_id:
        format: uuid
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: không tìm thấy dữ liệu
        type: string
    type: object
  service.AuthResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/service.UserResponse'
    type: object
  service.DueVocabulary:
    properties:
      audio_url:
        type: string
      meaning:
        type: string
      next_review_at:
        type: string
      phonetic:
        type: string
      srs_stage:
        type: integer
      term:
        example: apple
        type: string
      vocabulary_id:
        format: uuid
        type: string
    type: object
  service.LanguageResponse:
    properties:
      code:
        example: en-US
        type: string
      created_at:
        type: string
      id:
        example: en
        type: string
      name:
        example: English
        type: string
    type: object
  service.LoginRequest:
    properties:
      identifier:
        example: user@laclingo.vn
        type: string
      password:
        example: matkhau123
        type: string
    type: object
  service.RegisterRequest:
    properties:
      email:
        example: user@laclingo.vn
        type: string
      full_name:
        example: Nguyễn Văn A
        maxLength: 100
        type: string
      password:
        example: matkhau123
        maxLength: 72
        minLength: 8
        type: string
      username:
        example: nguyenvana
        maxLength: 50
        minLength: 3
        type: string
    type: object
  service.UpdateProfileRequest:
    properties:
      avatar_url:
        example: https://example.com/avatar.png
        type: string
      full_name:
        example: Nguyễn Văn A
        maxLength: 100
        type: string
    type: object
  service.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        example: user@laclingo.vn
        type: string
      full_name:
        type: string
      id:
        format: uuid
        type: string
      streak_count:
        type: integer
      username:
        example: nguyenvana
        type: string
    type: object
info:
  contact: {}
  description: API học ngôn ngữ với thuật toán SRS (SM-2).
  title: LacLingo API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email hoặc username, và mật khẩu
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Đăng nhập
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Mật khẩu 8-72 ký tự. Trả về access token để đăng nhập luôn.
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Đăng ký tài khoản
      tags:
      - auth
  /languages:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.LanguageResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Danh sách ngôn ngữ
      tags:
      - languages
  /languages/{id}:
    get:
      parameters:
      - description: Language ID
        example: en
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LanguageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Lấy thông tin ngôn ngữ
      tags:
      - languages
  /srs/due:
    get:
      parameters:
      - description: Số lượng tối đa (mặc định 20, tối đa 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.DueVocabulary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Từ vựng đến hạn ôn tập
      tags:
      - srs
  /srs/reviews:
    post:
      consumes:
      - application/json
      parameters:
      - description: Kết quả ôn tập
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VocabularyReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VocabularyReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ghi nhận kết quả ôn tập (SM-2)
      tags:
      - srs
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Danh sách user
      tags:
      - users
  /users/{id}:
    get:
      parameters:
      - description: User ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin user
      tags:
      - users
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thông tin user đang đăng nhập
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Chỉ cập nhật các field được gửi lên.
      parameters:
      - description: Thông tin cần cập nhật
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật hồ sơ của user đang đăng nhập
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Nhập "Bearer <access_token>" (lấy từ /auth/login hoặc /auth/register)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"